refactor!: modernize registry interface
