+++ src/OrasProject.Oras/Registry/IManifestStore.cs
@@ -13,13 +13,12 @@
 
 using OrasProject.Oras.Content;
 
-namespace OrasProject.Oras.Interfaces.Registry
+namespace OrasProject.Oras.Registry;
+
+/// <summary>
+/// IManifestStore is a CAS with the ability to stat and delete its content.
+/// Besides, IManifestStore provides reference tagging.
+/// </summary>
+public interface IManifestStore : IBlobStore, IReferencePushable, ITaggable
 {
-    /// <summary>
-    /// IManifestStore is a CAS with the ability to stat and delete its content.
-    /// Besides, IManifestStore provides reference tagging.
-    /// </summary>
-    public interface IManifestStore : IBlobStore, IReferencePusher, ITaggable
-    {
-    }
 }
