+++ src/OrasProject.Oras/Registry/Remote/Auth/HttpClientWithBasicAuth.cs
