+++ src/OrasProject.Oras/Registry/Remote/LinkUtility.cs
