+++ src/OrasProject.Oras/Registry/Remote/RemoteReference.cs
