+++ src/OrasProject.Oras/Registry/Remote/ResponseTypes.cs
